Change the order of buildplan candidates such that it goes from most specific to least specific
This changes the order of the buildplans generated at detection. It orders them from most specific to least specific. The rationale behind this is that the lifecycle will process buildplans in depth-first fashion and it will stop when it finds the first successful buildplan candidate (successful meaning all the provides/requires are met for all passing buildpacks). Previously, you could get into a condition where a less specific candidate buildpack would be selected over a more specific selection just because the less specific candidate was listed first. Changing the order will ensure that the most specific plan is picked first, which is the desired behavior.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>