Fixes bug where JRE was being installed into the launch container for native-image apps
The test cases were out of sync with the buildplans being retuned from detect, so we missed a case where jre, jdk and native-image-builder were present in the build plan. This triggers a condition where the JRE is installed when it should not be. This PR adds a test to cover this properly & fixes the logic so the JRE is not installed for native-image apps.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>