Add ARM64 dependencies
Adds the JRE, JDK and NIK dependencies for ARM64.

This commit also adds update workflows to keep these dependencies up-to-date. This will require https://github.com/paketo-buildpacks/pipeline-builder/pull/1348 to be merged first, because the workflows for ARM64 need to be able to pass the arch through to the GH Action that checks for Bellsoft updates.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>